Major changes restructure complete, issues with testing
