+++ exporter/SynthesisFusionAddin/src/global_manager.py
