+++ exporter/SynthesisFusionAddin/src/parser/synthesis_parser/physical_properties.py
