+++ exporter/SynthesisFusionAddin/src/parser/synthesis_parser/utilities.py
