+++ exporter/SynthesisFusionAddin/src/ui/command_groups/joint_settings_command_group.py
@@ -2,8 +2,8 @@ from enum import Enum
 from ...general_imports import *
 from ...configure import NOTIFIED, write_configuration
 from ...Analytics.alert import showAnalyticsAlert
-from .. import Helper, FileDialogConfig, OsHelper, CustomGraphics, IconPaths
-from ...Parser.ParseOptions import (
+from .. import helper, file_dialog_config, os_helper, graphics_custom, icon_paths
+from ...parser.parse_options import (
     Gamepiece,
     Mode,
     ParseOptions,
@@ -12,7 +12,7 @@ from ...Parser.ParseOptions import (
     JointParentType,
 )
 
-from ..Configuration.SerialCommand import SerialCommand
+from ..configuration.serial_command import SerialCommand
 
 class JointSettingsCommandGroup(MenuItem):
